moved configuration constants to config module
