+++ src/goose/config/__init__.py
@@ -1,6 +1,7 @@
 from pathlib import Path
+from typing import Final
 
-GOOSE_GLOBAL_PATH = Path("~/.config/goose").expanduser()
-PROFILES_CONFIG_PATH = GOOSE_GLOBAL_PATH.joinpath("profiles.yaml")
-SESSIONS_PATH = GOOSE_GLOBAL_PATH.joinpath("sessions")
-SESSION_FILE_SUFFIX = ".jsonl"
+GOOSE_GLOBAL_PATH: Final = Path("~/.config/goose").expanduser()
+PROFILES_CONFIG_PATH: Final = GOOSE_GLOBAL_PATH.joinpath("profiles.yaml")
+SESSIONS_PATH: Final = GOOSE_GLOBAL_PATH.joinpath("sessions")
+SESSION_FILE_SUFFIX: Final = ".jsonl"
